v0.1.2 - Support for wrapping errors
Changed Err.Expect(msg)
from panic(msg)
to panic(&Err[T]{fmt.Errorf("%s: %w", msg, self.Error)})
So Err.Expect(msg) will wrap the underlying error with msg. Expect()
will now be caught using Catch(). This is a common practice in Golang
and now this can be done easily with this library. Added a section in
the README for the same.

As a consequence Expect() for Option and ExpectErr() for Result have
been removed. It could end up being confusing as there is no error
wrapping required in those calls and so the given string is useless.